# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/confirm:
    post:
      operationId: confirmBilling
      tags: [billing]
      summary: Refresh billing status
      description: Forces a Stripe sync and reconcile. Used after returning from Checkout, since the managed webhook only fires against the production domain.
      responses:
        "200":
          description: Billing refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingConfirmResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /attorneys/{id}/billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [billing]
      summary: Start a subscription checkout for a listing
      description: Creates a Stripe Checkout session to subscribe an approved listing you own, which makes it publicly visible.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionInput"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Listing not eligible for checkout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /attorneys/{id}/billing/portal:
    post:
      operationId: createBillingPortalSession
      tags: [billing]
      summary: Open the billing portal for a listing
      description: Creates a Stripe billing portal session to manage or cancel a listing's subscription.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portal session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPortalSession"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: No billing account yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /attorneys/{id}/billing/upgrade-pro:
    post:
      operationId: upgradeListingToPro
      tags: [billing]
      summary: Upgrade a subscribed listing to the Pro tier
      description: Switches an existing active subscription you own to the Pro plan (prorated). The listing must already be subscribed; new listings subscribe via checkout, then upgrade here.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription upgraded to Pro
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingConfirmResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Listing not eligible for upgrade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Pro plan not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs. These can optionally
        be protected with authentication or ACL checks based on the use case.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CreateCheckoutSessionInput:
      type: object
      properties:
        flow:
          type: string
          enum: [signup, dashboard]
          description: Which page started checkout; controls where Stripe redirects back to. The plan tier is auto-selected server-side (Founding while available, otherwise Basic).
    CheckoutSession:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: Stripe Checkout URL to redirect the attorney to.
    BillingPortalSession:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: Stripe billing portal URL.
    BillingConfirmResult:
      type: object
      required:
        - ok
        - updated
      properties:
        ok:
          type: boolean
        updated:
          type: integer
          description: Number of listings whose billing columns were updated.
...